requirements.txt
streamlit
numpy
matplotlib
pillow
requests
plotly
scikit-learn